Do not add ssh port option when port == 22
Do not add the ssh -o Port option when the port is equal to the default
ssh port. As port may be specified in with an ssh config file (via the
-F option) any port specified in such a file will not be honored.

This patch allows for the port to be provided by an ssh configuration
file. As the DEFAULT_REMOTE_PORT is specified with a default integer of
22, the conditional "if self.port is not None" can never be reached by
normal means.